+++ lib/dito_sdk.dart
@@ -1,24 +1,24 @@
 library dito_sdk;
 
 import 'dart:convert';
-import 'package:dito_sdk/constants.dart';
-import 'package:dito_sdk/entity/user.dart';
-import 'package:dito_sdk/entity/event.dart';
-import 'package:dito_sdk/database.dart';
-import 'package:dito_sdk/services/firebase_messaging_service.dart';
-import 'package:dito_sdk/utils/http.dart';
-import 'package:dito_sdk/utils/sha1.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:http/http.dart' as http;
 
+import './constants.dart';
+import './utils/http.dart';
+import './utils/sha1.dart';
+import './services/notification_service.dart';
+import './database.dart';
+import './entity/event.dart';
+import './entity/user.dart';
+
 class DitoSDK {
   String? _apiKey;
   String? _secretKey;
-  String? _userID;
-  String? _signature;
-  User? _user;
+  late Map<String, String> _assign;
   late NotificationService _notificationService;
+  User _user = User();
 
   static final DitoSDK _instance = DitoSDK._internal();
 
@@ -32,15 +32,21 @@ class DitoSDK {
     return _notificationService;
   }
 
+  User get user {
+    return _user;
+  }
 
   void initialize({required String apiKey, required String secretKey}) async {
     _apiKey = apiKey;
     _secretKey = secretKey;
-    _signature = convertToSHA1(_secretKey!);
-    _notificationService = NotificationService(this);
+    _notificationService = NotificationService(_instance);
+    _assign = {
+      'platform_api_key': apiKey,
+      'sha1_signature': convertToSHA1(_secretKey!),
+    };
   }
 
-  Future<void> initializePushService() async {
+  Future<void> initializePushNotificationService() async {
     await Firebase.initializeApp();
     await _notificationService.initialize();
 
@@ -67,16 +73,14 @@ class DitoSDK {
 
     final params = {
       'id_type': 'id',
-      'platform_api_key': _apiKey,
-      'sha1_signature': _signature,
       'network_name': 'pt',
       'event': jsonEncode(event.toJson())
     };
 
-    final url =
-        Uri.parse(Constants.endpoints
-        .replace(value: _userID!, endpoint: Endpoint.events));
+    params.addAll(_assign);
 
+    final url = Uri.parse(Constants.endpoints
+        .replace(value: _user.id!, endpoint: Endpoint.events));
 
     return await Api().post(url, params);
   }
@@ -99,12 +103,9 @@ class DitoSDK {
   }
 
   Future<void> _setUserId(String userId) async {
-    _userID = userId;
-    _verifyPendingEvents();
-  }
-
-  User? get user {
-    return _user;
+    if (_user.isValid) {
+      _verifyPendingEvents();
+    }
   }
 
   void identify({
@@ -117,42 +118,34 @@ class DitoSDK {
     String? location,
     Map<String, String>? customData,
   }) {
-    if (_user == null) {
-      _user = User(
-          userID: userID,
-          name: name,
-          email: email,
-          gender: gender,
-          birthday: birthday,
-          location: location,
-          cpf: cpf,
-          customData: customData);
+    _user.userID = userID;
 
-      return;
+    if (cpf != null) {
+      _user.cpf = cpf;
     }
 
     if (name != null) {
-      _user?.name = name;
+      _user.name = name;
     }
 
     if (email != null) {
-      _user?.email = email;
+      _user.email = email;
     }
 
     if (gender != null) {
-      _user?.gender = gender;
+      _user.gender = gender;
     }
 
     if (birthday != null) {
-      _user?.birthday = birthday;
+      _user.birthday = birthday;
     }
 
     if (location != null) {
-      _user?.location = location;
+      _user.location = location;
     }
 
     if (customData != null) {
-      _user?.customData = customData;
+      _user.customData = customData;
     }
 
     _setUserId(userID);
@@ -161,8 +154,8 @@ class DitoSDK {
   Future<void> setUser(User user) async {
     _user = user;
 
-    if (_user!.valid) {
-      await _setUserId(_user!.id);
+    if (_user.isValid) {
+      await _setUserId(_user.id!);
     } else {
       throw Exception(
           'User registration is required. Please call the identify() method first.');
@@ -172,20 +165,19 @@ class DitoSDK {
   Future<http.Response> identifyUser() async {
     _checkConfiguration();
 
-    if (!_user!.valid) {
+    if (_user.isNotValid) {
       throw Exception(
           'User registration is required. Please call the identify() method first.');
     }
 
     final params = {
-      'platform_api_key': _apiKey,
-      'sha1_signature': _signature,
-      'user_data': jsonEncode(_user!.toJson()),
+      'user_data': jsonEncode(_user.toJson()),
     };
 
-    final url = Uri.parse(
-      Constants.endpoints
-        .replace(value: _user!.id, endpoint: Endpoint.identify));
+    params.addAll(_assign);
+
+    final url = Uri.parse(Constants.endpoints
+        .replace(value: _user.id!, endpoint: Endpoint.identify));
 
     return await Api().post(
       url,
@@ -209,7 +201,7 @@ class DitoSDK {
       customData: customData,
     );
 
-    if (_userID == null) {
+    if (_user.isNotValid) {
       final database = LocalDatabase.instance;
       await database.createEvent(event);
       return http.Response("", 200);
@@ -221,22 +213,46 @@ class DitoSDK {
   Future<http.Response> registryMobileToken({required String token}) async {
     _checkConfiguration();
 
-    if (_userID == null) {
+    if (_user.isNotValid) {
       throw Exception(
-          'User registration is required. Please call the setUserId() method first.');
+          'User registration is required. Please call the identify() method first.');
     }
 
     final params = {
       'id_type': 'id',
-      'platform_api_key': _apiKey,
-      'sha1_signature': _signature,
       'token': token,
       'platform': Constants.platform,
     };
 
-    final url = Uri.parse(
-        Constants.endpoints
-        .replace(value: _userID!, endpoint: Endpoint.registryMobileTokens));
+    params.addAll(_assign);
+
+    final url = Uri.parse(Constants.endpoints
+        .replace(value: _user.id!, endpoint: Endpoint.registryMobileTokens));
+
+    return await Api().post(
+      url,
+      params,
+    );
+  }
+
+  Future<http.Response> removeMobileToken({required String token}) async {
+    _checkConfiguration();
+
+    if (_user.isNotValid) {
+      throw Exception(
+          'User registration is required. Please call the identify() method first.');
+    }
+
+    final params = {
+      'id_type': 'id',
+      'token': token,
+      'platform': Constants.platform,
+    };
+
+    params.addAll(_assign);
+
+    final url = Uri.parse(Constants.endpoints
+        .replace(value: _user.id!, endpoint: Endpoint.removeMobileTokens));
 
     return await Api().post(
       url,
@@ -251,8 +267,6 @@ class DitoSDK {
     _checkConfiguration();
 
     final params = {
-      'platform_api_key': _apiKey,
-      'sha1_signature': _signature,
       'channel_type': 'mobile',
       'data': jsonEncode({
         'identifier': identifier,
@@ -260,8 +274,10 @@ class DitoSDK {
       })
     };
 
+    params.addAll(_assign);
+
     final url = Uri.parse(Constants.endpoints
-        .replace(value: _userID!, endpoint: Endpoint.openNotification));
+        .replace(value: _user.id!, endpoint: Endpoint.openNotification));
 
     return await Api().post(
       url,