+++ lib/entity/custom_notification.dart
@@ -1,5 +1,6 @@
-import 'package:dito_sdk/entity/data_payload.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
+
+import 'data_payload.dart';
 
 class CustomNotification {
   final int id;