+++ CHANGELOG.md
@@ -1,10 +1,13 @@
-## 1.0.0-beta-4 (2024-04-29)
+## 0.5.0 (2024-04-30)
 
 ### Novos Recursos
 
 - Adicionados os seguintes métodos:
   - `setUser()`: Método para salvar os dados do usuário antes de um identify.
-  - `initializePushService()`: Método para inicializar o serviço de mobile push notification.
+  - `removeMobileToken()`: Método para remover o token de um usuário.
+  - `initializePushNotificationService()`: Método para inicializar o serviço de mobile push notification.
+  - `setAndroidDetails()`: Método para personalizar o serviço de mobile push notification no Android.
+  - `setIosDetails()`: Método para personalizar o serviço de mobile push notification no iOS.
 
 ### Mudanças